Update Batocera-CRT-Script.sh
Small cosmetic changes to how Boot resolutions are displayed to the user during script setup. 

Explanation
Boot resolutions are always one pixel off (+1) this is to make sure that the geometry script (will be added soon) can work correctly. 
This has no effect for the user and nothing that will ever be noticed to the naked eye.