Update batocera-resolution-v38_MYZAR_ZFEbHVUE
For ceration of Modeline with switchres using Dotclock_min 0.

In setMode because for intel (VGA) when we use standalone with for example 640x480; swithres doesn't able to find the modeline if dotclock_min is equal to 25.0. so put dotclock = 0. make it the modeline possible for game.

In defineMode because when we want to use 640x480 or 768x576i (Because we can use any modelines in the menu/developper under ES) for ES we cannot use it with dotclock = 25.0. 

So for these cases with forced dotclock_min =0 create the modeline for game or ES and put again dotclock_min = 25 after.