VNC Server Scaling Script
There are built in commands for scaling mention in the man pages for **x11vnc** but they work really lousy and make everything really slow.
Since the only thing important for us is to scale the xfce window then we can use xrandr transform instead. 
Emulation Station will still be in your boot resolution i.e. 640x480 

But as default it's setup to use 1080p and that is more then enough.