Update Batocera-CRT-Script-v41.sh
Add in two scripts at the start

* Multiple monitor connection detection
The Script will exit if two or more outputs are connected during script setup.

* 10-monitor detection
The Script will detect & delete the file
``/etc/X11/xorg.conf.d/10-monitor.conf`` 
If it is found and ask the user to reboot. 
This way we always make sure that the correct outputs are disabled and also if the user made a mistake it is easily fixed by running the script again.

Sure it adds one more layer of complexity but better that then a broken setup where files needs to be manually deleted.  

Both of these will be at the very beginning. 
They are well explained and easy to see how they work
If there are issue with it it can easily be removed or disabled.